Add CSV target language
Prior to this change, there were two "target languages" for the
license report output: Markdown and HTML. Both output formats are
optimized for human consumption.

This change adds a third target language: CSV (comma-separated
values). The CSV version of the license report is a machine-readable
data file that is useful when users want the license report output
to become the input for another tool.