Remove unused property from TargetLanguage
TargetLanguage is a trait representing an output format for the
license report. It has a property called header4 which is not used
when the license report is generated. If header4 ever were used, it
would be a problem for the CSV target language, so it's best to just
remove it entirely.